Complete v1 auth/users/admin runtime with unified API contract
