+++ README.md
@@ -105,11 +105,11 @@ Copy-Item .env.example .env
 
 $env:PYTHONPATH="src"
 
-python -m flask --app wsgi:app db upgrade
-python -m flask --app wsgi:app forge seed
-python -m flask --app wsgi:app forge create-admin --email admin@example.com --password Password123
+PYTHONPATH=src python -m flask --app flaskforge.wsgi:app db upgrade
+PYTHONPATH=src python -m flask --app flaskforge.wsgi:app forge seed
+PYTHONPATH=src python -m flask --app flaskforge.wsgi:app forge create-admin --email admin@example.com --password Password123
 
-python -m flask --app wsgi:app run --debug
+PYTHONPATH=src python -m flask --app flaskforge.wsgi:app run --debug
 ````
 
 Open:
@@ -131,11 +131,11 @@ python -m pip install -e ".[dev]"
 
 cp .env.example .env
 
-PYTHONPATH=src python -m flask --app wsgi:app db upgrade
-PYTHONPATH=src python -m flask --app wsgi:app forge seed
-PYTHONPATH=src python -m flask --app wsgi:app forge create-admin --email admin@example.com --password Password123
+PYTHONPATH=src python -m flask --app flaskforge.wsgi:app db upgrade
+PYTHONPATH=src python -m flask --app flaskforge.wsgi:app forge seed
+PYTHONPATH=src python -m flask --app flaskforge.wsgi:app forge create-admin --email admin@example.com --password Password123
 
-PYTHONPATH=src python -m flask --app wsgi:app run --debug
+PYTHONPATH=src python -m flask --app flaskforge.wsgi:app run --debug
 ```
 
 ---
