+++ docs/rbac.md
@@ -12,6 +12,13 @@
 - `users:write`
 - `roles:read`
 - `roles:write`
+- `permissions:read`
+- `permissions:write`
+
+## Enforcement
+
+- Regular users can only read/update their own user resource.
+- Admin users can manage users, roles, and permissions.
 
 ## Decorators
 
@@ -19,12 +26,3 @@
 - `@require_roles(*roles)`
 - `@require_permissions(*permissions)`
 - `@require_owner_or_permission(permission, owner_param="user_id")`
-
-## Admin role assignment example
-
-```bash
-curl -X POST http://127.0.0.1:5000/api/admin/users/2/roles \
-  -H "Authorization: Bearer <ADMIN_ACCESS_TOKEN>" \
-  -H "Content-Type: application/json" \
-  -d '{"name":"staff","action":"assign"}'
-```
