+++ scripts/init_db.sh
@@ -2,3 +2,4 @@
 set -euo pipefail
 
 PYTHONPATH=src flask --app wsgi:app db upgrade
+PYTHONPATH=src flask --app wsgi:app forge seed
