+++ src/api/v1/admin/__init__.py
@@ -1,0 +1,3 @@
+from .routes import admin_bp
+
+__all__ = ["admin_bp"]
