+++ src/api/v1/auth/__init__.py
@@ -1,0 +1,3 @@
+from .routes import auth_bp
+
+__all__ = ["auth_bp"]
