+++ src/api/v1/health.py
@@ -1,8 +1,10 @@
-from flask import Blueprint, jsonify
+from flask import Blueprint
+
+from core.responses import ok
 
 health_bp = Blueprint("health", __name__)
 
 
 @health_bp.get("/health")
 def health():
-    return jsonify({"status": "ok"})
+    return ok({"status": "ok"})
