+++ src/api/v1/users/__init__.py
@@ -1,0 +1,3 @@
+from .routes import users_bp
+
+__all__ = ["users_bp"]
