+++ src/flaskforge/__init__.py
@@ -1,0 +1,3 @@
+from app import create_app
+
+__all__ = ["create_app"]
